#1 Completed implementation

+++ backend/src/main/kotlin/org/hikit/er/controller/response/internal/LocalityInternalResponse.kt
@@ -3,4 +3,4 @@ package org.hikit.er.controller.response.internal
 import org.hikit.er.rest.LocalityDto
 
 data class LocalityInternalResponse (val data: List<LocalityDto>,
-                                     val totalCount: Long)
+                                     val totalCount: Int)
